+++ src/services/EventControllerAttendeeService.ts
@@ -20,12 +20,20 @@ export class EventControllerAttendeeService {
         return await this.ecaRepo.getController(eventId, userId);
     }
 
+    public async controllerAssignmentExists(eventId: number, userId: number) {
+        return await this.ecaRepo.controllerAssignmentExists(eventId, userId);
+    }
+
     public async createControllerAssignment(
         eventId: number,
         userId: number,
         airportId: number,
         positionId: number
     ) {
+        const exists = await this.controllerAssignmentExists(eventId, userId);
+
+        // Future implementation of the error class story... ref. User/Event services for more in depth explanations
+        if (exists) return false;
         return await this.ecaRepo.createControllerAssignment(eventId, userId, airportId, positionId);
     }
 
@@ -34,6 +42,10 @@ export class EventControllerAttendeeService {
         userId: number,
         change: IEventControllerAttendeeUpdateDto
     ) {
+        const exists = await this.controllerAssignmentExists(eventId, userId);
+
+        // Future implementation of the error class story... ref. User/Event services for more in depth explanations
+        if (!exists) return false;
         return await this.ecaRepo.updateControllerAssignment(eventId, userId, change);
     }
 
@@ -41,6 +53,10 @@ export class EventControllerAttendeeService {
         eventId: number,
         userId: number
     ) {
+        const exists = await this.controllerAssignmentExists(eventId, userId);
+
+        // Future implementation of the error class story... ref. User/Event services for more in depth explanations
+        if (!exists) return true;
         return await this.ecaRepo.deleteControllerAssignment(eventId, userId);
     }
 }
